Bump waitForExpectations timeout to work around test timing out
See http://disq.us/p/273n65m

At somepoint, between the time of writing and Xcode 11.2.1, something
changed in such a way that it now somehow takes more than one second for
the expectation to be fulfilled.

This is a **workaround**. I'll need to dig more into this to understand
what's going on. Is it an issue with my code or something messed up in
Xcode?